Introduce TaskFactory class
Now the Tasks are going to be created in the TaskFactory and
BlueprintService will only do the orchestration between reading and creating everything.

This will allow having more complex parsing in the future
